add loopback shunt detection
significant changes:
-no more software override for shunt detection. driver only supports
hardware shunt in place for new rule ingestion
-loopback check inits the LOOP A/B pins to in and out respectively,
then flips the out and checks for a change on in. reports success or
failure; used for security when deciding to ingest new rules in the filter

--tests have narrowed the window for pin change propagation to 3 microseconds,
and the function currently uses this delay twice, once after pin state init,
and once after pin state toggle before the check. Banks of 100000 tests
in this configuration have shown no errors in operation. Removing the init
each time allows for 1-2 microseconds wait only for toggle propagation,
but introduces a small but random error spread (approx 1/2500 rate, 0.04%)

--note: it may be desirable to improve security to introduce this check
right before rule assembly/storage to flash. this is beyond this change
and test scope of this commit