so far unsuccessful attempts to register pin change with external interrupt
not sure what is the issue:
pin changes correctly when mcp has an interrupt. eic does not seem to be
watching properly, if set to watch for low level, it will trigger infinitely
and be unable to clear. if set for anything else, it does not see the pin
change. this makes me suspect that it is not watching the correct pin at all,
but for the ivi for instance, pin 67 corresponds to the int pin and the gpio
port for eicint[1]...