test of firewall default block and modify. mostly successful.
this implementation puts a single applicable rule at the end of a set of
inapplicable rules, to ensure that the speed is worst case for the set.
this implementation is only using a set of 16 rules. the rule in this test
touches all bytes to attempt to ensure no access is optimized.